Got React app testing infrastructure up and running, minor refactors to nav bar code
