+++ .gitignore
@@ -1,7 +1,11 @@
 node_modules/
 .env
-.package-lock.json
+package-lock.json
 
 server/secrets.js
 public/bundle.js
 client/assets
+
+.DS_Store
+client/.DS_Store
+server/.DS_Store
