+++ jest-teardown.js
@@ -1,0 +1,3 @@
+module.exports = async (globalConfig) => {
+  testServer.close();
+};
