+++ package.json
@@ -7,9 +7,19 @@
 		"start": "nodemon server/server.js",
 		"build": "NODE_ENV=production webpack",
 		"dev": "concurrently \"NODE_ENV=development nodemon server/server.js\" \"NODE_ENV=development webpack serve --open --hot\"",
-		"test": "echo \"Error: no test specified\" && exit 1"
+		"test": "jest --verbose"
 	},
 	"author": "aecubed",
+	"jest": {
+		"testEnvironment": "jest-environment-jsdom",
+		"setupFilesAfterEnv": [
+			"@testing-library/jest-dom/extend-expect"
+		],
+		"verbose": true,
+		"moduleNameMapper": {
+			"^.+\\.(css|less|scss)$": "babel-jest"
+		}
+	},
 	"dependencies": {
 		"@babel/plugin-transform-async-to-generator": "^7.18.6",
 		"@babel/runtime": "^7.18.9",
@@ -38,7 +48,11 @@
 		"@babel/plugin-transform-runtime": "^7.18.9",
 		"@babel/preset-env": "^7.1.0",
 		"@babel/preset-react": "^7.0.0",
+		"@testing-library/jest-dom": "^5.16.4",
+		"@testing-library/react": "^12.1.5",
+		"@testing-library/user-event": "^14.4.1",
 		"autoprefixer": "^10.4.8",
+		"babel-jest": "^28.1.3",
 		"babel-loader": "^8.2.3",
 		"concurrently": "^5.0.0",
 		"cross-env": "^6.0.3",
@@ -47,10 +61,14 @@
 		"eslint-plugin-react": "^7.21.5",
 		"file-loader": "^6.2.0",
 		"html-webpack-plugin": "^5.5.0",
+		"jest": "^28.1.3",
+		"jest-environment-jsdom": "^28.1.3",
 		"nodemon": "^1.19.4",
 		"postcss": "^8.4.14",
 		"postcss-loader": "^7.0.1",
+		"react-test-renderer": "^16.14.0",
 		"style-loader": "^3.3.1",
+		"supertest": "^6.2.4",
 		"webpack": "^5.64.1",
 		"webpack-cli": "^4.10.0",
 		"webpack-dev-server": "^4.5.0",
